// integration-service/lib/zoominfo/types.ts
export type FirmographicResult = {
	zi_c_location_id: number;
	zi_c_company_id: number;
	zi_c_is_hq: "0" | "1";
	zi_es_ecid: number;
	zi_es_location_id: string;
	zi_c_tier_grade: string;
	zi_c_name: string;
	zi_c_name_display: string;
	zi_c_legal_entity_type: string | null;
	zi_c_url: string;
	zi_c_street: string;
	zi_c_street_2: string | null;
	zi_c_city: string;
	zi_c_state: string;
	zi_c_zip: string;
	zi_c_country: string;
	zi_c_cbsa_name: string | null;
	zi_c_county: string;
	zi_c_latitude: number;
	zi_c_longitude: number;
	zi_c_verified_address: "0" | "1";
	zi_c_employee_range: string;
	zi_c_employees: number | null;
	zi_c_revenue_range: string;
	zi_c_revenue: number | null;
	zi_c_phone: string | null;
	zi_c_fax: string | null;
	zi_c_industry_primary: string;
	zi_c_sub_industry_primary: string;
	zi_c_industries: string;
	zi_c_sub_industries: string;
	zi_es_industry: string;
	zi_es_industries_top3: string;
	zi_c_naics2: string;
	zi_c_naics4: string;
	zi_c_naics6: string;
	zi_c_naics_top3: string;
	zi_c_sic2: string;
	zi_c_sic3: string;
	zi_c_sic4: string;
	zi_c_sic_top3: string;
	zi_c_estimated_age: number | null;
	zi_c_year_founded: string | null;
	zi_c_is_b2b: "0" | "1";
	zi_c_is_b2c: "0" | "1";
	zi_es_hq_ecid: number;
	zi_es_hq_location_id: string;
	zi_c_company_name: string;
	zi_c_company_url: string;
	zi_c_company_street: string;
	zi_c_company_street_2: string | null;
	zi_c_company_city: string;
	zi_c_company_state: string;
	zi_c_company_zip: string;
	zi_c_company_country: string;
	zi_c_company_cbsa_name: string | null;
	zi_c_company_county: string;
	zi_c_company_latitude: number;
	zi_c_company_longitude: number;
	zi_c_company_verified_address: "0" | "1";
	zi_c_company_employee_range: string;
	zi_c_company_employees: number | null;
	zi_c_company_revenue_range: string;
	zi_c_company_revenue: number | null;
	zi_c_company_phone: string | null;
	zi_c_company_fax: string | null;
	zi_c_linkedin_url: string | null;
	zi_c_facebook_url: string | null;
	zi_c_twitter_url: string | null;
	zi_c_yelp_url: string | null;
	zi_c_domain_rank: number | null;
	zi_c_keywords: string;
	zi_c_top_keywords: string | null;
	zi_c_num_keywords: number | null;
	zi_c_employee_growth_1yr: number | null;
	zi_c_employee_growth_2yr: number | null;
	zi_es_growth: string | null;
	zi_es_employee_growth: string | null;
	zi_es_revenue_growth: string | null;
	zi_es_percent_employee_growth: number | null;
	zi_es_percent_revenue_growth: number | null;
	zi_c_name_confidence_score: number | null;
	zi_c_url_confidence_score: number | null;
	zi_c_address_confidence_score: number;
	zi_c_phone_confidence_score: number | null;
	zi_c_employees_confidence_score: number | null;
	zi_c_revenue_confidence_score: number | null;
	zi_es_industry_confidence_score: number | null;
	zi_c_naics_confidence_score: number | null;
	zi_c_sic_confidence_score: number | null;
	zi_es_industries_top3_confidence_scores: string;
	zi_c_naics_top3_confidence_scores: string;
	zi_c_sic_top3_confidence_scores: string;
	zi_c_ids_merged: string;
	zi_c_names_other: string | null;
	zi_c_url_status: string;
	zi_c_urls_alt: string | null;
	zi_c_url_last_updated: string | null;
	zi_c_inactive_flag: "0" | "1";
	zi_c_ein: string | null;
	zi_c_is_small_business: "0" | "1";
	zi_c_is_public: "0" | "1";
	zi_c_ticker: string | null;
	zi_c_tickers_alt: string | null;
	zi_c_has_mobile_app: "0" | "1";
	zi_c_currency_code: string;
	zi_c_num_locations: number;
	zi_c_hr_contacts: number | null;
	zi_c_sales_contacts: number | null;
	zi_c_marketing_contacts: number | null;
	zi_c_finance_contacts: number | null;
	zi_c_c_suite_contacts: number | null;
	zi_c_engineering_contacts: number | null;
	zi_c_it_contacts: number | null;
	zi_c_operations_contacts: number | null;
	zi_c_legal_contacts: number | null;
	zi_c_medical_contacts: number | null;
	zi_c_tech_ids: string;
	zi_c_latest_funding_age: number | null;
	zi_c_num_of_investors: number | null;
	zi_c_investor_names: string | null;
	zi_c_funding_strength: string | null;
	zi_c_funding_type: string | null;
	zi_c_total_funding_amount: number | null;
	zi_c_latest_funding_amount: number | null;
	zi_c_latest_funding_date: string | null;
	zi_c_num_funding_rounds: number | null;
	zi_c_is_fortune_100: "0" | "1";
	zi_c_is_fortune_500: "0" | "1";
	zi_c_is_s_and_p_500: "0" | "1";
	zi_c_last_updated_date: string;
	zi_c_release_date: string;
};

export type ZoomInfoResponse = {
	match: MatchResult;
	firmographic?: FirmographicResult;
	match_mode?: "heuristic" | "ai";
};

export type ZoomInfoEntityMatchTask = EntityMatchTask<{ firmographic?: FirmographicResult | null }>;